Removed references to component which is no longer supported by Google.
Quote: You are receiving this communication because you have one or more projects that have recently used the Dialogflow CX console.

We’re writing to let you know that access to the Dialogflow CX console will be removed for all users on October 31, 2025. Dialogflow CX console users will be routed to the new Conversational Agents console to consolidate the user experience.